backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3

SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

bcrypt==4.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9

python-socketio[asgi]==5.11.3

pydantic==2.7.4
python-jose==3.3.0